WIP: allocation for managed subjects
