+++ contracts/components/staking/IStakeSubject.sol
@@ -17,5 +17,6 @@ interface IStakeSubject {
 interface IDelegatedStakeSubject is IStakeSubject {
     function getTotalManagedSubjects(uint256 subject) external view returns(uint256);
     function getManagedStakeThreshold(uint256 managedId) external view returns(StakeThreshold memory);
+    function getManagerIdFor(uint256 managedSubjectId) external view returns(uint256);
 
 }
